# Conflicts:
#   BE/src/main/java/com/chibbol/wtz/global/timer/service/NewTimerService.java

+++ BE/src/main/java/com/chibbol/wtz/global/stomp/service/StompService.java
@@ -1,12 +1,13 @@
 package com.chibbol.wtz.global.stomp.service;
 
 import com.chibbol.wtz.domain.room.entity.Room;
-import com.chibbol.wtz.global.stomp.repository.StompRepository;
 import com.chibbol.wtz.domain.room.service.RoomEnterInfoRedisService;
 import com.chibbol.wtz.domain.room.service.RoomService;
 import com.chibbol.wtz.domain.user.entity.User;
+import com.chibbol.wtz.domain.user.repository.UserRepository;
 import com.chibbol.wtz.domain.user.service.UserService;
 import com.chibbol.wtz.global.stomp.dto.DataDTO;
+import com.chibbol.wtz.global.stomp.repository.StompRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.listener.ChannelTopic;
@@ -23,6 +24,7 @@ public class StompService {
     private final ChannelTopic roomTopic;
     private final RoomEnterInfoRedisService roomEnterInfoRedisService;
     private final RoomService roomService;
+    private final UserRepository userRepository;
 
     public void connectUser(String sessionId, Long userSeq) {
         log.info("connectUser 시작");
@@ -37,8 +39,15 @@ public class StompService {
     }
 
     public void unsubscribeUser(Long userSeq) {
+        if(userSeq == null) {
+            return;
+        }
+
         log.info("EXIT 시작");
         String roomCode = stompRepository.getRoomCodeByUserSeq(userSeq);
+        if(roomCode == null) {
+            return;
+        }
         User user = userService.findByUserSeq(userSeq);
         // 메세지 보내기
         DataDTO dataDTO = DataDTO.builder()
@@ -75,18 +84,37 @@ public class StompService {
     }
 
     public void disconnectUser(String sessionId) {
+        if(sessionId == null) {
+            return;
+        }
+
         log.info("disconnect 시작");
         Long userSeq = stompRepository.getUserSeqBySessionId(sessionId);
+        if(userSeq == null) {
+            return;
+        }
         if (stompRepository.checkExistRoom(userSeq)) {
             log.info("이게 왜 안뜨지?");
             unsubscribeUser(userSeq);
         }
         stompRepository.removeUserSeqForSessionId(sessionId);
         stompRepository.removeSessionIdForUserSeq(userSeq);
+        // 로그아웃 처리
+        User user = userRepository.findByUserSeq(userSeq);
+        if(user == null) {
+            return;
+        }
+        user.updateRefreshToken(null);
+        userRepository.save(user);
+
         log.info("disconnect 끝");
     }
 
     public boolean checkForDuplicateUser(Long userSeq) {
+        if(userSeq == null) {
+            return false;
+        }
+
         return stompRepository.checkForDuplicateUser(userSeq);
     }
 }
